OPZ editor and separate code to specific files
Also separate wavetable synth tab code into separate file
